thêm api xóa favorite bằng favoriteId

+++ src/main/java/com/lowquality/serverwebm/service/FavoriteService.java
@@ -21,6 +21,8 @@ public class FavoriteService {
     private MangaService mangaService;
     @Autowired
     private UserService userService;
+    @Autowired
+    private PermissionService permissionService;
 
     public List<FavoriteDTO> findByUserId(Integer userId) {
         return favoriteRepository.findByUser_Id(userId).stream().map(this::convertToDTO).collect(Collectors.toList());
@@ -30,11 +32,11 @@ public class FavoriteService {
         return favoriteRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Favorite not found"));
     }
-    public FavoriteDTO addFavorite(Integer userId,Integer mangaId) {
-        if (favoriteRepository.findByUser_IdAndManga_Id(userId,mangaId)!=null) {
+    public FavoriteDTO addFavorite(Integer mangaId) {
+        User user= permissionService.getCurrentUser();
+        if (favoriteRepository.findByUser_IdAndManga_Id(user.getId(),mangaId)!=null) {
             throw new IllegalArgumentException("Manga đã nằm trong danh sách yêu thích");
         }
-        User user = userService.findById(userId);
     Mangadetail manga= mangaService.getMangaEntityById(mangaId);
     Favorite favorite = new Favorite();
     favorite.setUser(user);
@@ -44,6 +46,8 @@ public class FavoriteService {
     }
     public void removeFavorite(Integer favoriteID) {
         Favorite favorite = findById(favoriteID);
+        User user = userService.findById(favorite.getUser().getId());
+        permissionService.checkUserPermission(user.getId(),"thay đổi favorite");
         favoriteRepository.delete(favorite);
     }
 
@@ -55,8 +59,11 @@ public class FavoriteService {
                 .build();
     }
 
-    public void removeFavoriteByUserAndManga(Integer userId, Integer mangaId) {
-        Favorite favorite = favoriteRepository.findByUser_IdAndManga_Id(userId,mangaId);
+    public void removeFavoriteByUserAndManga( Integer mangaId) {
+        User user= permissionService.getCurrentUser();
+        Favorite favorite = favoriteRepository.findByUser_IdAndManga_Id(user.getId(), mangaId);
+
+        permissionService.checkUserPermission(user.getId(),"thay đổi favorite");
         favoriteRepository.delete(favorite);
     }
 }
